addressed some PR comments
