+++ python/tank/descriptor/descriptor_core.py
@@ -1,15 +1,17 @@
 # Copyright (c) 2016 Shotgun Software Inc.
-# 
+#
 # CONFIDENTIAL AND PROPRIETARY
-# 
-# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit 
+#
+# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit
 # Source Code License included in this distribution package. See LICENSE.
-# By accessing, using, copying or modifying this work you indicate your 
-# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights 
+# By accessing, using, copying or modifying this work you indicate your
+# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
 # not expressly granted therein are reserved by Shotgun Software Inc.
 
 from .descriptor import Descriptor
+from .errors import TankMissingManifestError
 from . import constants
+
 
 class CoreDescriptor(Descriptor):
     """
@@ -42,3 +44,37 @@ class CoreDescriptor(Descriptor):
         constraints["min_sg"] = manifest.get("requires_shotgun_version", constants.LOWEST_SHOTGUN_VERSION)
 
         return constraints
+
+    def get_feature_info(self, feature_name, default_value=None):
+        """
+        Retrieves information for a given feature in the manifest.
+
+        The ``default_value`` will be returned in the following cases:
+            - a feature is missing from the manifest
+            - the manifest is empty
+            - the manifest is missing
+
+        :param str feature_name: Name of the feature to retrieve from the manifest.
+        :param obj default_value: Value to return if the feature is missing.
+
+        :returns: The value for the feature if present, ``default_value`` otherwise.
+        """
+        infos = self.get_features_info()
+        if feature_name in infos:
+            return infos[feature_name]
+        else:
+            return default_value
+
+    def get_features_info(self):
+        """
+        Retrieves the feature dictionary from the manifest.
+
+        If the manifest if empty or missing, an empty dictionary will be returned.
+
+        :returns: Dictionary of features.
+        """
+        try:
+            manifest = self._get_manifest() or {}
+        except TankMissingManifestError:
+            return {}
+        return manifest.get("features") or {}
